將 roomAsList() 移到 ShowRoom.java 內
