+++ GoogleSheet/src/main/java/us/dontcareabout/googleSheet/Exhibition.java
@@ -3,7 +3,6 @@ package us.dontcareabout.googleSheet;
 import us.dontcareabout.googleSheet.Exceptions.DateIntervalErrorException;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -17,14 +16,14 @@ public class Exhibition {
 	public Exhibition(RawData data) {
 		this.name = data.name;
 		this.displayDate = new DateInterval(data.start, data.end);
-		rooms.addAll(roomAsList(data.rooms));
+		rooms.addAll(ShowRoom.roomAsList(data.rooms));
 	}
 
 	/**
 	 * 新增換展資訊
 	 */
 	public void addChangeInfo(RawData data) {
-		for (String r : roomAsList(data.rooms)) {
+		for (String r : ShowRoom.roomAsList(data.rooms)) {
 			DateInterval changeInterval = new DateInterval(data.start, data.end);
 			if (!displayDate.containInterval(changeInterval)) {
 				throw new DateIntervalErrorException(changeInterval + " is not between " + displayDate);
@@ -33,23 +32,6 @@ public class Exhibition {
 		}
 	}
 
-	/**
-	 * 將展廳資料轉為 List，並以半個展廳為最小單位。
-	 */
-	public static List<String> roomAsList(String rooms) {
-		List<String> showRooms = new ArrayList<String>();
-
-		for (String r : rooms.split(",")) {
-
-			char subRoom = r.charAt(r.length() - 1);
-			if (subRoom == 'A' || subRoom == 'B') {
-				showRooms.add(r);
-				break;
-			}
-			showRooms.addAll(Arrays.asList(r + "A", r + "B"));
-		}
-		return showRooms;
-	}
 
 	@Override
 	public String toString() {
